pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grid-congestion"
version = "0.1.0"
description = "Analyze congestion patterns in PJM and ERCOT using nodal electricity price data."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "gridstatus>=0.35.0",
  "numpy>=1.26",
  "pandas>=2.2",
  "plotly>=5.24",
  "PyYAML>=6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
